significant rewrite
1. separate composability rules by execution strategy.

I am not sure if these should be separate in the final document, but separating helps me reason about them better within this draft stage.

2. remove discussion of runtime translation from composability rules

thats basically mixing the idea of "transforms" with composability.

For example, if there is a way of translating a missing enum value for a given subschema into an enum value that is present, we are basically transforming that subschema into a subschema that has the missing enum value by using a transformRequest method

Instead, let's make the composability rules strict and discuss separately the idea of transforms