Add item to 11 Jan agenda
I thought it would be helpful to start an open discussion up front about the meaning of different terms being used in this space.  It would be great to hear from attendees on how they interpret the meaning of terms like "Composite Schemas," "Federation," and "Fusion" so we can use those names in ways that are clear to everyone.