Add Jeff's notes from 2024-01 primary meeting
I inadvertently took these notes aloof of the provided Google Doc link. Apologies for not using that doc in the first place!  These are largely duplicative of the notes above but there are some details covered above that weren't covered here, and some detail here that's not covered above, so I figured it was worth adding these in case it's helpful